# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Portal API",
        "description": "API untuk aplikasi Job Portal - Pertemuan 6 PPW2",
        "contact": {
            "name": "Akmal Manggala Putra",
            "email": "akmalmanggala2810@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/public/jobs": {
            "get": {
                "tags": [
                    "Public Jobs"
                ],
                "summary": "Get public job listings (No authentication required)",
                "description": "Mendapatkan daftar lowongan pekerjaan publik tanpa memerlukan token autentikasi. Endpoint read-only untuk akses publik.",
                "operationId": "9e527edeaad8ee56050c51edbef6f348",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by title, company, or description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Filter by location",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "company",
                        "in": "query",
                        "description": "Filter by company name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "job_type",
                        "in": "query",
                        "description": "Filter by job type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Full Time",
                                "Part Time",
                                "Contract",
                                "Freelance"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "example": 2
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Software Engineer"
                                                    },
                                                    "company": {
                                                        "type": "string",
                                                        "example": "Tech Corp"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Jakarta"
                                                    },
                                                    "salary": {
                                                        "type": "string",
                                                        "example": "10000000-15000000"
                                                    },
                                                    "job_type": {
                                                        "type": "string",
                                                        "example": "Full Time"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Job description..."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/public/jobs/{id}": {
            "get": {
                "tags": [
                    "Public Jobs"
                ],
                "summary": "Get public job detail (No authentication required)",
                "description": "Mendapatkan detail lowongan pekerjaan tanpa memerlukan token autentikasi",
                "operationId": "5db67ef8871de8831dc6da69a069391b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Software Engineer"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Job description..."
                                        },
                                        "company": {
                                            "type": "string",
                                            "example": "Tech Corp"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "Jakarta"
                                        },
                                        "salary": {
                                            "type": "string",
                                            "example": "10000000-15000000"
                                        },
                                        "job_type": {
                                            "type": "string",
                                            "example": "Full Time"
                                        },
                                        "logo": {
                                            "type": "string",
                                            "example": "logos/company-logo.png"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}